move TEXT_FEATURE_MAPPING into constants.py
